test: extract 18ntext matchter
and prepped addtl test case
